Finish lab3 with full points.
NOTES: Due to the mistake in merging the original repo, the grading
function of Lab3 is broken. But if you test the code in a new cloned
repo, it can pass the test and get full points.

Try `cd Lab3 && make submit` to see my submission and copy them to a new
repo.